Worked on current actor marker changing when characters added or removed.
